Cleanup plex config flow tests (#88991)
